requirements.txt
# AI Prompt Engineer Competition Requirements
# Python Version: 3.12+ (recommended: 3.12.8 or 3.13.6)
# Core AI/ML Libraries
openai>=1.0.0
litellm>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Text Processing and NLP
#nltk>=3.8.0
#spacy>=3.6.0
#transformers>=4.30.0
torch>=2.0.0,<2.5.0
#tokenizers>=0.13.0

# Web Scraping and API Tools
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0

# Data Storage and Serialization
# sqlite3 is built into Python, no need to install
json5>=0.9.0
pyyaml>=6.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter and Notebook Support
#jupyter>=1.0.0
#ipykernel>=6.25.0
#notebook>=6.5.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6  
rich>=13.0.0
click>=8.1.0

# Optional: Advanced AI Libraries
# langchain>=0.0.200
# anthropic>=0.3.0
# cohere>=0.0.0

# DOCX support
python-docx>=1.1.0

# Environment variables
python-dotenv>=1.0.0
python-pptx>=0.6.21
